Don't pass area when not using areas
When area is passed when we're not using areas, generated urls in child actions will have `?area=` appended to them. This is caused by the following:

`ChildActionExtensions.ActionHelper` [calls][0] `RouteCollectionExtensions.GetVirtualPathForArea` which [loops over][1] the configured routes and sets `usingAreas` to `true` if there's a route found with an area. Then back in the `ActionHelper` it will [remove the `area` from the route values][2] when `usingAreas` was set to `true`.

So when we're not using areas, `ActionHelper` will pass the key `area` with value `""` in the `RouteData` to `Server.Execute`.

Finally when in `UrlHelper.GenerateUrl` it [merges the route values from the request][3] and the resulting `mergedRouteValues` will contain a key `area`. This is then [passed to `RouteCollectionExtensions.GetVirtualPathForArea`][4] which will [remove the key `area` but only if (again) `usingAreas` is `true`][5].

[0]: https://github.com/aspnet/AspNetWebStack/blob/ba26cfbfbf958d548e4c0a96e853250f13450dc6/src/System.Web.Mvc/Html/ChildActionExtensions.cs#L108
[1]: https://github.com/aspnet/AspNetWebStack/blob/ba26cfbfbf958d548e4c0a96e853250f13450dc6/src/System.Web.Mvc/RouteCollectionExtensions.cs#L41
[2]: https://github.com/aspnet/AspNetWebStack/blob/ba26cfbfbf958d548e4c0a96e853250f13450dc6/src/System.Web.Mvc/Html/ChildActionExtensions.cs#L119
[3]: https://github.com/aspnet/AspNetWebStack/blob/ba26cfbfbf958d548e4c0a96e853250f13450dc6/src/System.Web.Mvc/UrlHelper.cs#L188
[4]: https://github.com/aspnet/AspNetWebStack/blob/ba26cfbfbf958d548e4c0a96e853250f13450dc6/src/System.Web.Mvc/UrlHelper.cs#L190
[5]: https://github.com/aspnet/AspNetWebStack/blob/ba26cfbfbf958d548e4c0a96e853250f13450dc6/src/System.Web.Mvc/RouteCollectionExtensions.cs#L102